requirements.txt
Flask==2.0.1
gunicorn==20.1.0
pdfplumber==0.7.4
Pillow==9.5.0
pytesseract==0.3.10
python-dotenv==0.19.2
openai==1.3.0
langchain==0.0.340
langchain-community==0.0.10
faiss-cpu==1.7.4
pandas==1.5.3
Werkzeug==2.2.3